FIXED normalize_gray_levels() function in pre_processing_functions.py
The previous version of this function, introduced in commit 3296eca, did not take into account that the image passed as argument might have only zeros.
With this commit that bug is fixed by checking that at least one element of the image must be different from zero, otherwise the image itself is returned.