Implemented UNet model and unit testing
This commit introduces the class UNet, responsible for the creation of the model,
made by the components introduced in 'U_Net_components'.
The architecture of the model is the following:

- Encoder component: Performs 4 consecutive 'DownSample' operations, each reducing the
                    dimension of the input by half while doubling the number of channels (feature maps);
- Bottle Neck: Performs a 'DoubleConvolution' operation on the output of the encoder, doubling the channels;
- Decoder component: Performs 4 consecutive 'UpSample' operation with skip connections, restoring
                    the original dimension of the image and reducing the number of channels to 64;
- Output layer: Performs a single convolution with kernel_size = 1, returning the output of the segmentation.

The proper dimension of the output has been tested in 'test_unet_components'.
Dropout layers will be added to the model in order to make it an approximated bayesian U-Net.