Changed data root directory structure and dataset class
This commit introduce significant canges in the way the data is organised before being passed as input to the training loop.
Now the dataset class expects the data root directory to be organised as follows:
./"data directory"
    ./"FLAIR"
    ./"LESION"
Where the FLAIR subdirectory contains all the axial slices obtained from the 3D images of all the patients, and the LESION subdirectory
contains the corresponding lesion masks with the same index. Only FLAIR scans for which the corresponding lesion mask contains at least
a lesion pixel are stored.

List of the main changes in each method:
-__init__: Initialize the class object with scan_indexes, root_path as well as FLAIR and LESION paths
-__len__: Returns the number of FLAIR scans (data samples) instead of number of patients
-__getitem__:  now takes as argument an index and outputs a tuple containing a single FLAIR scan and the relative LESION
mask instead of a dictionary containing all FLAIR scans and LESION masks for a given patient.
-filter_slices: REMOVED. The filtering operation is now performed as a preliminar operation in the data directory preparation.

WHY THE CHANGE?
The changes were necessary because the previous dataset proved to be too complex to be used in combination with PyTorch DataLoaders.