Implemented predict script
This commit introduces predict.py, a script for obtaining predicted segmentation masks.
The prediction function accepts the input image, segmentation masks, model path, and device
to use as arguments. Post-prediction, pixels with values greater than 0.5 are set to 1, as
these values represent the probability of a pixel being part of the lesion class.
For comparison, the original image, prediction, and ground truth are plotted side by side.