Fix Focal loss formula and implemented unit testing
This commit aims to fix the formula for computing Focal loss.
It could not be computed by using F.binary_cross_entropy_with_logits, since it expects the input to be a tensor of un-normalized scores.
In the case of this architecture, the inputs are normalized. Therefore, F.binary_cross_entropy was employd.

Additionally, this commit introduces the file test_focal_loss.py, containing the following tests:
- Testing the proper initialization of the FocalLoss
- Tests ensure expected output shapes and types for different reduction modes.
- Manually calculated loss values are compared to the FocalLoss output for multiple scenarios (using mean reduction).